refact: move PDF rendering to own component
